# src/Presentation.Web.Server/wwwroot/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API",
    "version": "v1"
  },
  "paths": {
    "/api/tenants/{tenantId}/catalog/customers/{id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogCustomer",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Catalog"
        ],
        "operationId": "UpdateCatalogCustomer",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CustomerModel"
                  }
                ]
              }
            }
          },
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalog"
        ],
        "operationId": "DeleteCatalogCustomer",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/catalog/customers": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogCustomers",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "operationId": "CreateCatalogCustomer",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CustomerModel"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tenants/{tenantId}/catalog/categories/{id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogCategory",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryModel"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/catalog/categories/{id}/books": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogCategoryBooks",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/catalog/categories": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogCategories",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CustomerModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "PersonName": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PersonFormalNameModel"
              }
            ]
          },
          "Address": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/AddressModel"
              }
            ]
          },
          "Email": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PersonFormalNameModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Parts": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "Suffix": {
            "type": "string",
            "nullable": true
          },
          "Full": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AddressModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Line1": {
            "type": "string",
            "nullable": true
          },
          "Line2": {
            "type": "string",
            "nullable": true
          },
          "PostalCode": {
            "type": "string",
            "nullable": true
          },
          "City": {
            "type": "string",
            "nullable": true
          },
          "Country": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TagModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Edition": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "Sku": {
            "type": "string",
            "nullable": true
          },
          "Isbn": {
            "type": "string",
            "nullable": true
          },
          "Price": {
            "type": "number",
            "format": "decimal"
          },
          "Publisher": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/BookPublisherModel"
              }
            ]
          },
          "PublishedDate": {
            "type": "string",
            "format": "date"
          },
          "Keywords": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "Authors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BookAuthorModel"
            }
          },
          "Categories": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BookCategoryModel"
            }
          },
          "Chapters": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BookChapterModel"
            }
          },
          "Tags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TagModel"
            }
          },
          "Version": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookPublisherModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookAuthorModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Position": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BookCategoryModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookChapterModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Number": {
            "type": "integer",
            "format": "int32"
          },
          "Content": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CategoryModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Order": {
            "type": "integer",
            "format": "int32"
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "ParentId": {
            "type": "string",
            "nullable": true
          },
          "Children": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CategoryModel"
            }
          },
          "Version": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "oauth2",
        "description": "Oidc Authentication",
        "flows": {
          "implicit": {
            "authorizationUrl": "/protocol/openid-connect/auth",
            "tokenUrl": "/protocol/openid-connect/token"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}